requirements.txt
altair==5.0.1
astor==0.8.1
attrs==23.1.0
backports.zoneinfo==0.2.1
base58==2.1.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==7.1.2
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
easy-ocr==0.1.1
easyocr==1.7.0
filelock==3.12.2
fonttools==4.41.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
imageio==2.31.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
lazy_loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
mysql-connector-python==8.0.33
networkx==3.1
ninja==1.11.1
numpy==1.24.4
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==12.0.1
pyclipper==1.3.0.post4
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
rich==13.4.2
rpds-py==0.8.10
scikit-image==0.19.3
scipy==1.10.1
shapely==2.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.24.1
streamlit-option-menu==0.3.6
sympy==1.12
tenacity==8.2.2
tifffile==2023.7.10
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.6.0+cpu
tornado==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
zipp==3.16.2
